pyproject.toml
[project]
name = "rsa-toolbox"
version = "0.1.0"
description = "An experimental practicing project. A pack of various RSA commands."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pycryptodome>=3.23.0",
]
[[tool.uv.index]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
